narrow the recompose
